pyproject.toml
[project]
name = "cloudappsproject1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.2",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
]
